+++ tests/Integration/PlasterTemplates/Sampler/SimpleModulePlasterTemplate.Integration.Tests.ps1
@@ -79,6 +79,7 @@ Describe 'Simple Module Plaster Template' {
                 ModuleDescription = 'Module description'
                 ModuleVersion     = '1.0.0'
                 CustomRepo        = 'PSGallery'
+                MainGitBranch     = 'main'
             }
 
             { Invoke-Plaster @invokePlasterParameters } | Should -Not -Throw