+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added -WhatIf parameter to `release.module.build.ps1`
 
+### Changed
+
+- Updated Sampler documentation in README.md.
+
 ## [0.111.8] - 2021-08-08
 
 ### Changed