+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed `PesterOutputFormat` parameter in `DeployAll.PSDeploy.build.ps1`
  fix ([issue #292](https://github.com/gaelcolas/Sampler/issues/292)).
+- Removed the template `newDscCommunity` which is replaced by the template
+  `dsccommunity`.
 
 ### Added
 
@@ -18,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Merged both templates `dsccommunity` and `newDscCommunity` into the
+  template `dsccommunity`.
+- All templates now defaults to using `main` as the default branch.
 - Updated Sampler documentation in README.md.
 
 ## [0.111.8] - 2021-08-08